requirements.txt
delta-spark
delta
pyspark
bs4
azure-mgmt-resource 
azure-mgmt-datalake-store
azure-datalake-store
python-dotenv
tqdm
numpy